requirements.txt
ansicon==1.89.0
arrow==1.2.2
asgiref==3.5.2
async-timeout==4.0.2
blessed==1.19.1
cffi==1.15.0
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0.5
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.13.1
gunicorn==20.1.0
jinxed==1.2.0
mysqlclient==2.1.1
packaging==21.3
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
wcwidth==0.2.5
wrapt==1.14.1
